Updated event export to handle CURRENT X STATE entries.
The event and log classes were updated to detect and export
the CURRENT HOST STATE and CURRENT SERVICE STATE log entries.
Entries with an UP/OK state are exported as events with a
duration of zero. This ensures that all host/service objects
are tracked and the timestamp of their latest entry recorded.